pyproject.toml
[project]
name = "spoo-me"
version = "1.0.0"
description = "Open-Source URL Shortener Built with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.5",
    "crawlerdetect>=0.3.0",
    "dicttoxml>=1.7.16",
    "emoji>=2.14.1",
    "fastapi>=0.115.0",
    "geoip2>=5.1.0",
    "httpx>=0.28.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.0",
    "openpyxl>=3.1.5",
    "pycountry>=24.6.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.8.0",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "regex>=2026.1.15",
    "sentry-sdk>=2.44.0",
    "slowapi>=0.1.9",
    "structlog>=25.5.0",
    "tldextract>=5.3.0",
    "ua-parser[regex]>=1.0.1",
    "uvicorn[standard]>=0.34.0",
    "validators>=0.35.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
addopts = [
    "--import-mode=importlib",
    "-n", "auto",
    "--cov=.",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-config=pyproject.toml",
    "-p", "no:randomly",
]

[tool.coverage.run]
source = ["."]
parallel = true
omit = [
    "tests/*",
    "*/test_*",
    "*/conftest.py",
    ".venv/*",
    "main.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "pass",
    "\\.\\.\\.",
]
fail_under = 0
show_missing = true
skip_covered = true

[dependency-groups]
dev = [
    "anyio[trio]>=4.0.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.0",
    "pytest-clarity>=1.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.1",
    "pytest-randomly>=4.0.1",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.8.0",
    "ruff>=0.11.11",
    "uv>=0.9.6",
]
